Handle failures in ActionMailer that may raise exceptions
Failures from ActionMailer may prevent all details for the outgoing
request from getting filled in resulting in missing information
required to build the Agency and request info pages.

Treat these as a new class of send error, and show users that the
delivery of their messages failed, and administrator attention is
required.